pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.17.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.9",
    "streamlit-folium>=0.23.0",
    "streamlit>=1.39.0",
]
